backend/requirements.txt
Flask==2.2.2
PyPDF2
Werkzeug==2.2.2
boto3
botocore==1.37.21 
flask-cors
flask-jwt-extended
fpdf
livekit
livekit-agents
livekit-plugins-deepgram
livekit-plugins-openai
mariadb
numpy
openai
pymupdf
python-docx
python-dotenv
requests
whisper
